+++ app/model/console/spec_belong.mdl.php
@@ -86,35 +86,43 @@ class Spec_Belong extends Spec_Belong_Base {
 
 
 
-    function clear($num_no, $num_except = 0, $arr_search = array()) {
+    function clear($pagination = 0, $arr_search = array()) {
         $_arr_belongSelect = array(
             'belong_id',
             'belong_spec_id',
             'belong_article_id',
         );
 
-        $_arr_where = $this->queryProcess($arr_search);
+        $_arr_where         = $this->queryProcess($arr_search);
+        $_arr_pagination    = $this->paginationProcess($pagination);
+        $_arr_getData       = $this->where($_arr_where)->order('belong_id', 'DESC')->limit($_arr_pagination['limit'], $_arr_pagination['length'])->paginate($_arr_pagination['perpage'], $_arr_pagination['current'])->select($_arr_belongSelect);
 
-        $_arr_belongRows = $this->where($_arr_where)->order('belong_id', 'DESC')->limit($num_except, $num_no)->select($_arr_belongSelect);
+        if (isset($_arr_getData['dataRows'])) {
+            $_arr_clearData = $_arr_getData['dataRows'];
+        } else {
+            $_arr_clearData = $_arr_getData;
+        }
 
-        $_mdl_article = Loader::model('article');
-        $_mdl_spec    = Loader::model('spec');
+        if (!Func::isEmpty($_arr_clearData)) {
+            $_mdl_article = Loader::model('article');
+            $_mdl_spec    = Loader::model('spec');
 
-        foreach ($_arr_belongRows as $_key=>$_value) {
-            $_arr_articleRow = $_mdl_article->check($_value['belong_article_id']);
+            foreach ($_arr_clearData as $_key=>$_value) {
+                $_arr_articleRow = $_mdl_article->check($_value['belong_article_id']);
 
-            if ($_arr_articleRow['rcode'] != 'y120102') {
-                $this->delete(0, 0, false, false, false, false, $_value['belong_id']);
-            }
+                if ($_arr_articleRow['rcode'] != 'y120102') {
+                    $this->delete(0, 0, false, false, false, false, $_value['belong_id']);
+                }
 
-            $_arr_specRow = $_mdl_spec->check($_value['belong_spec_id']);
+                $_arr_specRow = $_mdl_spec->check($_value['belong_spec_id']);
 
-            if ($_arr_specRow['rcode'] != 'y180102') {
-                $this->delete(0, 0, false, false, false, false, $_value['belong_id']);
+                if ($_arr_specRow['rcode'] != 'y180102') {
+                    $this->delete(0, 0, false, false, false, false, $_value['belong_id']);
+                }
             }
         }
 
-        return $_arr_belongRows;
+        return $_arr_getData;
     }
